move utxo getAddress to UTXOBaseAdapter
